requirements.txt
torch>=1.12.0
torchvision>=0.13.0
pandas>=1.3.0
numpy>=1.21.0
Pillow>=8.0.0
tqdm>=4.60.0
scikit-learn>=0.24.0
